requirements.txt
gunicorn==19.9.0
python-dotenv==0.19.2
vk-api==11.9.6
